+++ examples/node-cache/CHANGELOG.md
@@ -1,5 +1,12 @@
 # cache-example
 
+## 0.1.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @express-route-cache/core@1.2.0
+
 ## 0.1.1
 
 ### Patch Changes